feat: style & minor edits
