+++ src/main/java/ru/practicum/shareit/booking/dto/BookingResponseDTO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.shareit.booking.entity.BookingStatus;
+import ru.practicum.shareit.booking.model.BookingStatus;
 import ru.practicum.shareit.item.projections.ItemBooking;
 import ru.practicum.shareit.user.projections.UserBooker;
