+++ src/main/java/ru/practicum/shareit/booking/model/Booking.java
@@ -1,12 +1,12 @@
-package ru.practicum.shareit.booking.entity;
+package ru.practicum.shareit.booking.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import ru.practicum.shareit.item.entity.Item;
-import ru.practicum.shareit.user.entity.User;
+import ru.practicum.shareit.item.model.Item;
+import ru.practicum.shareit.user.model.User;
 
 import javax.persistence.Id;
 import javax.persistence.Entity;
