+++ src/main/java/ru/practicum/shareit/constants/NamesJsonResponse.java
@@ -4,7 +4,7 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class NamesJsonResponse {
-    public static final String ERROR_MESSAGE = "Error message";
+    public static final String ERROR_MESSAGE = "errorMessage";
 
     public static final String ERROR_M_VALIDATION = "Validation error";
 }
