--- src/main/java/ru/practicum/shareit/item/projections/ItemIdWithBookingQueue.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.item.projections;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import ru.practicum.shareit.booking.dto.BookingQueueDTO;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class ItemIdWithBookingQueue {
-    private Long id;
-
-    private BookingQueueDTO lastBooking;
-
-    private BookingQueueDTO nextBooking;
-}
