+++ src/main/java/ru/practicum/shareit/item/repository/CommentRepository.java
@@ -1,15 +1,15 @@
-package ru.practicum.shareit.item.dao;
+package ru.practicum.shareit.item.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Component;
-import ru.practicum.shareit.item.entity.Comment;
-import ru.practicum.shareit.item.entity.Item;
+import ru.practicum.shareit.item.model.Comment;
+import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
 
 @Component
-public interface CommentDao extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
     @Query("SELECT com " +
             " FROM Comment AS com " +
             " JOIN FETCH com.item AS it " +
