+++ src/main/java/ru/practicum/shareit/request/entity/ItemRequest.java
@@ -2,7 +2,7 @@ package ru.practicum.shareit.request.entity;
 
 import lombok.Builder;
 import lombok.Data;
-import ru.practicum.shareit.user.entity.User;
+import ru.practicum.shareit.user.model.User;
 
 import java.time.LocalDateTime;
