+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user.entity;
+package ru.practicum.shareit.user.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
