+++ src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -1,12 +1,12 @@
-package ru.practicum.shareit.user.dao;
+package ru.practicum.shareit.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
-import ru.practicum.shareit.user.entity.User;
+import ru.practicum.shareit.user.model.User;
 
 import java.util.Optional;
 
 @Component
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findUserByEmailContainingIgnoreCase(String emailSearch);
 }
